backend/requirements.txt
# -------------------------
# Core Backend (FastAPI)
# -------------------------
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.1
python-dotenv==1.0.1

# -------------------------
# AI / LLM (Gemini)
# -------------------------
google-genai==0.3.0

# -------------------------
# RAG & Embeddings
# -------------------------
langchain==0.1.16
langchain-community==0.0.34
sentence-transformers==2.6.1
faiss-cpu==1.13.1

# -------------------------
# Data Processing & Scraping
# -------------------------
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1

# -------------------------
# Utilities
# -------------------------
tqdm==4.66.2
